+++ scheduler-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/scheduler/client/SchedulerServiceAsync.java
@@ -413,7 +413,7 @@ public interface SchedulerServiceAsync {
     /**
      *
      * @param sessionId current session
-     * @param signal signals that is to be sent to the job
+     * @param signal the signal that is to be sent to the job
      * @param jobId id of the job
      * @param asyncCallback the object used for notifying the caller when the asynchronous call is completed
      */
